Reformat helper to specific classes
